🔧 MAINT: Add errors for pandas 2.X for dataframes containing timestamps (#18)
pandas 2.X does not correctly handle pd.Timestamp in roundtrips to hdf5 files. The returned datatype is a datetime64[ns] instead of the expected datetime64[s]. Since everything else works with pandas 2 though, only a error message is added for dataframes created with these datatypes is added and the version requirements are not changed. In pandas 3.0 this issue seems to be already fixed.

Additionally some changes/updates to the CI are included (python versions, testing different pandas versions, etc.)